refactor: convert `apps/meteor/app/metrics/server/lib/statsTracker.js` to TypeScript
